Move admin/ to admin/service/
This is the first of a sequence of commits designed to isolate the
service and REST API code. This moves all of the code to service first
because the majority of code is service code and subsequent commits will
be easier to understand.
