+++ libsplinter/src/admin/service/shared.rs
